repo-tools: refactor root api report generation
