+++ packages/repo-tools/src/commands/api-reports/index.ts
@@ -14,7 +14,4 @@
  * limitations under the License.
  */
 
-export { runApiExtraction } from './runApiExtraction';
-export { buildDocs } from './buildDocs';
-export { createTemporaryTsConfig } from './createTemporaryTsConfig';
-export { generateTypeDeclarations } from './generateTypeDeclarations';
+export { buildApiReports } from './buildApiReports';
